+++ src/AddToForm.jsx
@@ -1,0 +1,17 @@
+
+
+const AddTodoForm = () => {
+
+    return(
+        
+        <form>
+            <label htmlFor="todoTitle"></label>
+            <input type="text" id="todoTitle"/> 
+            <button type="submit">Add</button>
+        </form>
+        
+    )
+
+}
+
+export default AddTodoForm
